mutants/pyproject.toml
[project]
name = "md-evals"
version = "0.1.0"
description = "Lightweight CLI tool for evaluating AI skills (SKILL.md) with Control vs Treatment testing using LiteLLM"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Javier Zader", email = "javier@example.com" }
]
keywords = ["ai", "llm", "evaluation", "skills", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "litellm>=1.40.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
]

[project.scripts]
md-evals = "md_evals.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.typer]
pretty_width = 100

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
